bug 1216786 - Make cascade deletion tests generic
Move identical cascade deletion test code to a generic TestCase, and run
API-specific tests for each namespace.
