+++ webplatformcompat/tests/test_viewsets.py
@@ -9,9 +9,75 @@ from rest_framework.test import APIRequestFactory
 from webplatformcompat.view_serializers import (
     ViewFeatureListSerializer, ViewFeatureRowChildrenSerializer,
     ViewFeatureSerializer)
-from webplatformcompat.models import Feature
+from webplatformcompat.models import (
+    Browser, Feature, Maturity, Section, Specification, Support, Version)
 
 from .base import APITestCase
+
+
+class TestCascadeDeleteGeneric(APITestCase):
+    __test__ = False  # Don't test against an unversioned API
+
+    def setUp(self):
+        super(TestCascadeDeleteGeneric, self).setUp()
+        self.login_user(groups=['change-resource', 'delete-resource'])
+        self.parent = self.create(Feature, slug='parent')
+        self.child = self.create(Feature, slug='doomed', parent=self.parent)
+        self.browser = self.create(Browser, slug='browser')
+        self.version = self.create(
+            Version, version='1.0', browser=self.browser)
+        self.support = self.create(
+            Support, version=self.version, feature=self.child)
+        self.maturity = self.create(Maturity, slug='MAT')
+        self.specification = self.create(
+            Specification, slug='Spec', maturity=self.maturity)
+        self.section = self.create(
+            Section, specification=self.specification)
+        self.section.features.add(self.child)
+
+    def assert_counts_after_delete(
+            self, url, browsers=1, versions=1, supports=1, maturities=1,
+            specifications=1, sections=1, features=2):
+        response = self.client.delete(url)
+        self.assertEqual(204, response.status_code)
+        self.assertEqual(browsers, Browser.objects.count())
+        self.assertEqual(versions, Version.objects.count())
+        self.assertEqual(supports, Support.objects.count())
+        self.assertEqual(maturities, Maturity.objects.count())
+        self.assertEqual(specifications, Specification.objects.count())
+        self.assertEqual(sections, Section.objects.count())
+        self.assertEqual(features, Feature.objects.count())
+
+    def test_delete_browser(self):
+        url = self.api_reverse('browser-detail', pk=self.browser.pk)
+        self.assert_counts_after_delete(
+            url, browsers=0, versions=0, supports=0)
+
+    def test_delete_version(self):
+        url = self.api_reverse('version-detail', pk=self.version.pk)
+        self.assert_counts_after_delete(url, versions=0, supports=0)
+
+    def test_delete_support(self):
+        url = self.api_reverse('support-detail', pk=self.support.pk)
+        self.assert_counts_after_delete(url, supports=0)
+
+    def test_delete_maturity(self):
+        url = self.api_reverse('maturity-detail', pk=self.maturity.pk)
+        self.assert_counts_after_delete(
+            url, maturities=0, specifications=0, sections=0)
+
+    def test_delete_specification(self):
+        url = self.api_reverse(
+            'specification-detail', pk=self.specification.pk)
+        self.assert_counts_after_delete(url, specifications=0, sections=0)
+
+    def test_delete_section(self):
+        url = self.api_reverse('section-detail', pk=self.section.pk)
+        self.assert_counts_after_delete(url, sections=0)
+
+    def test_delete_feature(self):
+        url = self.api_reverse('feature-detail', pk=self.parent.pk)
+        self.assert_counts_after_delete(url, features=0, supports=0)
 
 
 class TestUserBaseViewset(APITestCase):
